wh0x1s/requirements.txt
Flask==3.1.0
python-whois==0.9.5
validators==0.34.0

